[ERM-6204] Отслеживаем появление дублей при копировании прайса и позиции прайса

+++ BLCore/Source/API/2Gis.Erm.BLCore.API.Operations/Concrete/Prices/ICopyPricePositionOperationService.cs
@@ -5,6 +5,6 @@ namespace DoubleGis.Erm.BLCore.API.Operations.Concrete.Prices
 {
     public interface ICopyPricePositionOperationService : IOperation<CopyPricePositionIdentity>
     {
-        int Copy(long priceId, long sourcePricePositionId, long positionId);
+        long Copy(long priceId, long sourcePricePositionId, long positionId);
     }
 }
